// backend/src/modules/customer-order/enums/customer-order-status.enum.ts
export enum CUSTOMER_ORDER_STATUS {
  Draft = 'customerOrder.status.draft',
  Created = 'customerOrder.status.created',
  Validated = 'customerOrder.status.validated',
  Cancelled = 'customerOrder.status.cancelled',
  Invoiced = 'customerOrder.status.invoiced',
  Delivered = 'customerOrder.status.delivered',
}